Rename refactored component files in assets/js/components to use CamelCasing.
