+++ assets/js/components/ModuleIcon.js
