+++ assets/js/components/SettingsNotice.js
