+++ assets/js/components/dashboard-details/DashboardDetailsEntityView.js
@@ -28,7 +28,7 @@ import Data from 'googlesitekit-data';
 import { decodeHTMLEntity } from '../../util';
 import Link from '../Link';
 import WidgetContextRenderer from '../../googlesitekit/widgets/components/WidgetContextRenderer';
-import DateRangeSelector from '../date-range-selector';
+import DateRangeSelector from '../DateRangeSelector';
 import DashboardDetailsModules from './dashboard-details-modules';
 import DashboardDetailsEntityHeaderContainer from './DashboardDetailsEntityHeaderContainer';
 import { STORE_NAME as CORE_SITE } from '../../googlesitekit/datastore/site/constants';
