+++ assets/js/components/dashboard-details/dashboard-details-app.js
@@ -26,10 +26,10 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import Header from '../header';
+import Header from '../Header';
 import Link from '../Link';
-import PageHeader from '../page-header';
-import HelpLink from '../help-link';
+import PageHeader from '../PageHeader';
+import HelpLink from '../HelpLink';
 import DashboardDetailsEntityView from './DashboardDetailsEntityView';
 import DashboardDetailsEntityNotFoundView from './DashboardDetailsEntityNotFoundView';
 import { STORE_NAME as CORE_SITE } from '../../googlesitekit/datastore/site/constants';
