+++ assets/js/components/dashboard-splash/dashboard-splash-intro.js
@@ -31,7 +31,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Button from '../button';
-import Logo from '../logo';
+import Logo from '../Logo';
 
 class DashboardSplashIntro extends Component {
 	render() {
