+++ assets/js/components/data-block.js
@@ -31,7 +31,7 @@ import { sprintf } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import ChangeArrow from './change-arrow';
+import ChangeArrow from './ChangeArrow';
 import { numberFormat } from '../util/i18n';
 import SourceLink from './SourceLink';
