+++ assets/js/components/notifications/notification.js
@@ -38,7 +38,7 @@ import Button from '../button';
 import Warning from '../notifications/warning';
 import Error from '../notifications/error';
 import Link from '../Link';
-import ModuleIcon from '../module-icon';
+import ModuleIcon from '../ModuleIcon';
 
 class Notification extends Component {
 	constructor( props ) {
