+++ assets/js/components/setup-wizard/wizard-step-authentication.js
@@ -34,7 +34,7 @@ import { trackEvent } from '../../util';
 import Button from '../button';
 import Link from '../Link';
 import OptIn from '../optin';
-import HelpLink from '../help-link';
+import HelpLink from '../HelpLink';
 
 class WizardStepAuthentication extends Component {
 	render() {
