+++ assets/js/components/setup/search-console.js
@@ -33,7 +33,7 @@ import { Component, Fragment } from '@wordpress/element';
 import { trackEvent } from '../../util';
 import data, { TYPE_MODULES } from '../../components/data';
 import ProgressBar from '../ProgressBar';
-import HelpLink from '../../components/help-link';
+import HelpLink from '../HelpLink';
 import { Select, TextField, Input } from '../../material-components';
 import Button from '../../components/button';
