+++ assets/js/components/setup/setup-wrapper.js
@@ -32,9 +32,9 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import Header from '../header';
+import Header from '../Header';
 import Link from '../Link';
-import HelpLink from '../help-link';
+import HelpLink from '../HelpLink';
 import { getSiteKitAdminURL } from '../../util';
 import { STORE_NAME as CORE_SITE } from '../../googlesitekit/datastore/site/constants';
 import { STORE_NAME as CORE_MODULES } from '../../googlesitekit/modules/datastore/constants';
