+++ assets/js/modules/adsense/components/settings/SettingsView.js
@@ -25,7 +25,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import DisplaySetting from '../../../../components/display-setting';
+import DisplaySetting from '../../../../components/DisplaySetting';
 import Link from '../../../../components/Link';
 import { STORE_NAME } from '../../datastore/constants';
 import {
