+++ assets/js/modules/analytics/components/common/ExistingTagError.js
@@ -26,7 +26,7 @@ import { __, sprintf } from '@wordpress/i18n';
  */
 import Data from 'googlesitekit-data';
 import { STORE_NAME } from '../../datastore/constants';
-import ErrorText from '../../../../components/error-text';
+import ErrorText from '../../../../components/ErrorText';
 const { useSelect } = Data;
 
 export default function ExistingTagError() {
