+++ assets/js/modules/optimize/components/setup/SetupForm.js
@@ -35,7 +35,7 @@ import Button from '../../../../components/button';
 import { STORE_NAME } from '../../datastore/constants';
 import { isValidOptimizeID } from '../../util';
 import StoreErrorNotices from '../../../../components/StoreErrorNotices';
-import ErrorText from '../../../../components/error-text';
+import ErrorText from '../../../../components/ErrorText';
 import {
 	AMPExperimentJSONField,
 	OptimizeIDField,
