+++ stories/header.stories.js
@@ -5,7 +5,7 @@ import { storiesOf } from '@storybook/react';
 /**
  * Internal dependencies
  */
-import Header from '../assets/js/components/header';
+import Header from '../assets/js/components/Header';
 import { STORE_NAME as CORE_SITE } from '../assets/js/googlesitekit/datastore/site/constants';
 import { STORE_NAME as CORE_USER } from '../assets/js/googlesitekit/datastore/user/constants';
 import { WithTestRegistry } from '../tests/js/utils';
